+++ integrated_channels/utils.py
@@ -15,6 +15,7 @@ from urllib.parse import urlparse
 import pytz
 import requests
 
+from django.db.models import Q
 from django.utils import timezone
 from django.utils.html import strip_tags
 
@@ -381,3 +382,26 @@ def is_valid_url(url):
         return all([result.scheme, result.netloc])
     except ValueError:
         return False
+
+
+def batch_by_pk(ModelClass, extra_filter=Q(), batch_size=10000):
+    """
+    yield per batch efficiently
+    using limit/offset does a lot of table scanning to reach higher offsets
+    this scanning can be slow on very large tables
+    if you order by pk, you can use the pk as a pivot rather than offset
+    this utilizes the index, which is faster than scanning to reach offset
+    Example usage:
+    csod_only_filter = Q(integrated_channel_code='CSOD')
+    for items_batch in batch_by_pk(ContentMetadataItemTransmission, extra_filter=csod_only_filter):
+        for item in items_batch:
+            ...
+    """
+    qs = ModelClass.objects.filter(extra_filter).order_by('pk')[:batch_size]
+    while qs.exists():
+        yield qs
+        # qs.last() doesn't work here because we've already sliced
+        # loop through so we eventually grab the last one
+        for item in qs:
+            start_pk = item.pk
+        qs = ModelClass.objects.filter(pk__gt=start_pk).filter(extra_filter).order_by('pk')[:batch_size]